CreationTime refactoring Implementation moved from database to object model (data import should not change the migrated creation time)
