+++ CommonConcepts/Plugins/Rhetos.Dom.DefaultConcepts/Rhetos.Dom.DefaultConcepts.csproj
@@ -86,6 +86,7 @@
     <Compile Include="Computations\ComputeForNewBaseItemsCodeGenerator.cs" />
     <Compile Include="Computations\QueryableExtensionCodeGenerator.cs" />
     <Compile Include="SimpleBusinessLogic\ActionCodeGenerator.cs" />
+    <Compile Include="SimpleBusinessLogic\CreationTimeCodeGenerator.cs" />
     <Compile Include="SimpleBusinessLogic\DeactivatableCodeGenerator.cs" />
     <Compile Include="SimpleBusinessLogic\ReferenceCascadeDeleteCodeGenerator.cs" />
     <Compile Include="Computations\ExtensionComputedFromCodeGenerator.cs" />
